+++ lib/set/exception.ex
@@ -1,0 +1,11 @@
+defmodule Unicode.Set.ParseError do
+  @moduledoc """
+  Exception raised when an a Unicode Set cannot
+  be parsed
+  """
+  defexception [:message]
+
+  def exception(message) do
+    %__MODULE__{message: message}
+  end
+end